2 Example Projects for Eval Kit
Adding 2 example projects with instructions for eval03 usage.

"Example-NeoPixel-RSSI" reads the RSSI value from the Tile every 5 seconds and changes the color of the RGB LED in the eval kit

"Example-DHT22" reads the temperature (°C) and humidity (%) from a DHT22 sensor every 30 minutes. The data is formatted to a $TD command and sent to the Tile to queue for transmission. Also performs an RSSI measurement and changes the color of the LED